+++ src/Modules/Expirator/Controllers/BlockEditorController.php
@@ -150,6 +150,8 @@ class BlockEditorController implements InitializableInterface
             $metaboxTitle = $settingsFacade->getMetaboxTitle() ?? __('Future Actions', 'post-expirator');
             $metaboxCheckboxLabel = $settingsFacade->getMetaboxCheckboxLabel() ?? __('Enable Future Action', 'post-expirator');
 
+            $hiddenFields = (array) $this->hooks->applyFilters(HooksAbstract::FILTER_HIDDEN_METABOX_FIELDS, [], $post->post_type);
+
             wp_localize_script(
                 'postexpirator-block-editor',
                 'publishpressFutureBlockEditorConfig',
@@ -166,6 +168,7 @@ class BlockEditorController implements InitializableInterface
                     'taxonomyName' => $taxonomyPluralName,
                     'taxonomyTerms' => $taxonomyTerms,
                     'hideCalendarByDefault' => $settingsFacade->getHideCalendarByDefault(),
+                    'hiddenFields' => $hiddenFields,
                     'strings' => [
                         'category' => __('Categories', 'post-expirator'),
                         'panelTitle' => $metaboxTitle,