+++ src/Modules/Expirator/Controllers/ClassicEditorController.php
@@ -69,7 +69,8 @@ class ClassicEditorController implements InitializableInterface
 
         $this->hooks->addAction(
             CoreHooksAbstract::ACTION_SAVE_POST,
-            [$this, 'processMetaboxUpdate']
+            [$this, 'processMetaboxUpdate'],
+            20
         );
 
         $this->hooks->addAction(
@@ -92,12 +93,10 @@ class ClassicEditorController implements InitializableInterface
                 $id = $post->ID;
             } elseif (isset($post->post_id)) {
                 $id = $post->post_id;
+            } elseif (method_exists($post, 'get_id')) {
+                $id = $post->get_id();
             } elseif (isset($post->id)) {
                 $id = $post->id;
-            } else {
-                if (method_exists($post, 'get_id')) {
-                    $id = $post->get_id();
-                }
             }
 
             if (! is_null($id)) {
@@ -391,6 +390,8 @@ class ClassicEditorController implements InitializableInterface
             $metaboxTitle = $settingsFacade->getMetaboxTitle() ?? __('Future Actions', 'post-expirator');
             $metaboxCheckboxLabel = $settingsFacade->getMetaboxCheckboxLabel() ?? __('Enable Future Action', 'post-expirator');
 
+            $hiddenFields = (array) $this->hooks->applyFilters(HooksAbstract::FILTER_HIDDEN_METABOX_FIELDS, [], $postType);
+
             wp_localize_script(
                 'publishpress-future-classic-editor',
                 'publishpressFutureClassicEditorConfig',
@@ -408,6 +409,7 @@ class ClassicEditorController implements InitializableInterface
                     'postType' => $currentScreen->post_type,
                     'isNewPost' => $isNewPostPage,
                     'hideCalendarByDefault' => $settingsFacade->getHideCalendarByDefault(),
+                    'hiddenFields' => $hiddenFields,
                     'strings' => [
                         'category' => __('Category', 'post-expirator'),
                         'panelTitle' => $metaboxTitle,