+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnPostRowAction.php
@@ -64,7 +64,7 @@ class OnPostRowAction implements StepTypeInterface
     {
         return [
             [
-                "label" => __("Conditions", "post-expirator"),
+                "label" => __("Post Filter", "post-expirator"),
                 "description" => __(
                     "Specify the criteria for posts that will trigger this action.",
                     "post-expirator"